fix: bundle via vite for cjs support
