+++ package.json
@@ -15,28 +15,32 @@
   "type": "module",
   "exports": {
     ".": {
-      "import": "./dist/index.mjs",
+      "import": "./dist/index.js",
       "require": "./dist/index.cjs"
     }
   },
   "main": "dist/index.cjs",
-  "module": "dist/index.mjs",
-  "source": "src/index.mts",
+  "module": "dist/index.js",
+  "source": "src/index.ts",
   "files": [
     "src/**",
     "dist/**"
   ],
   "scripts": {
-    "build": "tsc -p ./src/tsconfig.esm.json && tsc -p ./src/tsconfig.json ",
+    "build": "run-p build.modern build.legacy build.tsc",
+    "build.legacy": "vite build --mode legacy",
+    "build.modern": "vite build --mode modern",
+    "build.tsc": "tsc --pretty -p tsconfig.build.json",
+    "dev": "vite build --watch",
     "format": "prettier . --write",
     "link.self": "pnpm link .",
-    "lint": "run-s link.self build && eslint . --fix",
+    "lint": "run-s link.self build && cross-env NODE_OPTIONS=--enable-source-maps eslint . --fix",
     "test": "run-s link.self build && run-p test.mjs test.cjs test.eslint8 test.format test.lint",
-    "test.cjs": "eslint spec/**/*.{yml,yaml} --config ./eslint.config.cjs",
-    "test.eslint8": "cross-env ESLINT_USE_FLAT_CONFIG=false npx -y eslint@8 spec/**/*.{yml,yaml} --config .eslintrc.cjs",
+    "test.cjs": "cross-env NODE_OPTIONS=--enable-source-maps eslint spec/**/*.{yml,yaml} --config ./eslint.config.cjs",
+    "test.eslint8": "cross-env NODE_OPTIONS=--enable-source-maps ESLINT_USE_FLAT_CONFIG=false npx -y eslint@8 spec/**/*.{yml,yaml} --config .eslintrc.cjs",
     "test.format": "prettier . --check",
     "test.lint": "eslint .",
-    "test.mjs": "eslint spec/**/*.{yml,yaml} --config ./eslint.config.mjs"
+    "test.mjs": "cross-env NODE_OPTIONS=--enable-source-maps eslint spec/**/*.{yml,yaml} --config ./eslint.config.mjs"
   },
   "dependencies": {
     "@types/estree": "^1.0.7",
@@ -46,18 +50,23 @@
     "jshint": "^2.13.6"
   },
   "devDependencies": {
+    "@babel/core": "^7.26.10",
+    "@types/babel__core": "~7.20.5",
     "@types/eslint": "8.56.10",
     "@types/js-yaml": "^4.0.9",
     "@types/jshint": "^2.12.4",
     "@typescript-eslint/eslint-plugin": "7.13.1",
     "@typescript-eslint/parser": "7.13.1",
+    "@upleveled/babel-plugin-remove-node-prefix": "^1.0.5",
     "cross-env": "^7.0.3",
     "eslint": "^9.5.0",
     "eslint-config-atomic": "^1.22.0",
     "npm-run-all2": "^6.2.0",
     "prettier": "^3.3.2",
     "prettier-config-atomic": "^4.0.0",
-    "typescript": "^5.4.5"
+    "typescript": "^5.4.5",
+    "vite": "^6.2.3",
+    "vite-plugin-babel": "^1.3.0"
   },
   "engines": {
     "node": ">=12"
