--- src/tsconfig.esm.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "module": "ESNext"
-  },
-  "include": ["./**/*.mts"]
-}
