--- src/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "../dist",
-    "rootDir": "./"
-  },
-  "include": ["./**/*.cts"]
-}
